requirements.txt
bcrypt
wget
python-dotenv
asyncio
mysql-connector-python
colorama
getpass